requirements.txt
streamlit
pandas
numpy
scikit-learn
matplotlib
requests
joblib
imbalanced-learn
fpdf2
plotly
xgboost
seaborn
